requirements.txt
streamlit
opencv-python-headless
torch
torchvision
numpy
pillow
matplotlib